Preserve more comments in template-literals
I found some cases where comments were being removed. Since I'd prefer
for this transform to be safe and produce less than ideal output, I
decided to just collect the comments and add them to the node. There are
some opportunities here to put them in better locations for better
formatting, but I'm going to leave that for another time.