Use the jscodeshift `use` api.
*This breaks this plugin's api* It also entirely removes the dependency on
jscodeshift itself. Instead, it exports a factory function that takes a config,
now also functions, and returns a `use`-able plugin to be registered with
jscodeshift.

As a result of moving away from the user invoking `registerMethods` directly,
we lose the ability to provide a custom name to the plugin's methods, but we
also avoid leaking the registration process to the user, and things are much
simpler IMHO.